refactor: avoid native control component frames
