+++ src/components/accordion/accordion.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { state } from '@askrjs/askr';
 import { Presence, Slot } from '@askrjs/askr/foundations/structures';
 import { composeRefs, mergeProps } from '@askrjs/askr/foundations/utilities';
@@ -353,9 +353,9 @@ export function AccordionTrigger(
   }
 
   return (
-    <NativeButton type={typeProp ?? 'button'} {...finalProps}>
+    <button type={typeProp ?? 'button'} {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
