+++ src/components/button/button.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { Slot } from '@askrjs/askr/foundations/structures';
 import { mergeProps } from '@askrjs/askr/foundations/utilities';
 import { getButtonInteractionProps } from './button-interactions';
@@ -99,8 +99,8 @@ export function Button(props: ButtonNativeProps | ButtonAsChildProps) {
   // Explicit type="button" default prevents accidental form submission
   const type = typeProp ?? 'button';
   return (
-    <NativeButton type={type} {...finalProps}>
+    <button type={type} {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
