+++ src/components/collapsible/collapsible.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { defineScope, readScope } from '@askrjs/askr';
 import { Slot, Presence } from '@askrjs/askr/foundations/structures';
 import { controllableState } from '@askrjs/askr/foundations/state';
@@ -145,9 +145,9 @@ export function CollapsibleTrigger(
   }
 
   return (
-    <NativeButton type="button" {...finalProps}>
+    <button type="button" {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
