+++ src/components/select/select-item.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { Slot } from '@askrjs/askr/foundations/structures';
 import { composeRefs, mergeProps } from '@askrjs/askr/foundations/utilities';
 import { pressable, rovingFocus } from '@askrjs/askr/foundations/interactions';
@@ -135,9 +135,9 @@ export function SelectItem(props: SelectItemProps | SelectItemAsChildProps) {
   }
 
   return (
-    <NativeButton type={type ?? 'button'} {...finalProps}>
+    <button type={type ?? 'button'} {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
