+++ src/components/select/select-trigger.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { Slot } from '@askrjs/askr/foundations/structures';
 import { composeRefs, mergeProps } from '@askrjs/askr/foundations/utilities';
 import { pressable } from '@askrjs/askr/foundations/interactions';
@@ -100,9 +100,9 @@ export function SelectTrigger(
   }
 
   return (
-    <NativeButton type={type ?? 'button'} {...finalProps}>
+    <button type={type ?? 'button'} {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
