+++ src/components/switch/switch.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { Slot } from '@askrjs/askr/foundations/structures';
 import { controllableState } from '@askrjs/askr/foundations/state';
 import { mergeProps } from '@askrjs/askr/foundations/utilities';
@@ -77,9 +77,9 @@ export function Switch(props: SwitchButtonProps | SwitchAsChildProps) {
   return (
     <>
       {hiddenInput}
-      <NativeButton type={typeProp ?? 'button'} {...finalProps}>
+      <button type={typeProp ?? 'button'} {...nativeButtonProps(finalProps)}>
         {children}
-      </NativeButton>
+      </button>
     </>
   );
 }
