+++ src/components/toggle-group/toggle-group-item.tsx
@@ -1,4 +1,4 @@
-import { NativeButton } from '../_internal/native-control';
+import { nativeButtonProps } from '../_internal/native-control';
 import { Slot } from '@askrjs/askr/foundations/structures';
 import { composeRefs, mergeProps } from '@askrjs/askr/foundations/utilities';
 import { pressable, rovingFocus } from '@askrjs/askr/foundations/interactions';
@@ -121,8 +121,8 @@ export function ToggleGroupItem(
   }
 
   return (
-    <NativeButton type={type ?? 'button'} {...finalProps}>
+    <button type={type ?? 'button'} {...nativeButtonProps(finalProps)}>
       {children}
-    </NativeButton>
+    </button>
   );
 }
